feat: add visual actions counter
